+++ CHANGELOG.md
@@ -7,3 +7,16 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Ingest new entities
+  - `esper_account`
+  - `esper_device`
+  - `esper_device_group`
+  - `esper_application`
+- Build new relationships
+  - `esper_account_has_device`
+  - `esper_account_has_device_group`
+  - `esper_account_has_application`
+  - `esper_device_group_has_device`